backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pytest==8.2.2
pytest-asyncio==0.23.7
ruff==0.4.10
